% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragraph2vec.R
\name{predict.paragraph2vec}
\alias{predict.paragraph2vec}
\title{Predict functionalities for a paragraph2vec model}
\usage{
\method{predict}{paragraph2vec}(
  object,
  newdata,
  type = c("nearest", "embedding"),
  top_n = 10L,
  encoding = "UTF-8",
  ...
)
}
\arguments{
\item{object}{a paragraph2vec model as returned by \code{\link{paragraph2vec}} or \code{\link{read.paragraph2vec}}}

\item{newdata}{for type 'embedding', \code{newdata} should be TODO}

\item{type}{either 'embedding' or 'nearest'. Defaults to 'nearest'.}

\item{top_n}{show only the top n nearest neighbours. Defaults to 10.}

\item{encoding}{set the encoding of the text elements to the specified encoding. Defaults to 'UTF-8'.}

\item{...}{not used}
}
\value{
depending on the type, you get a different result back:
\itemize{
\item{TODO}
\item{TODO}
}
}
\description{
Get either 
\itemize{
\item{the embedding of documents or words}
\item{TODO}
}
}
\examples{
## TODO
}
\seealso{
\code{\link{paragraph2vec}}, \code{\link{read.paragraph2vec}}
}
